refactor: 답글 api 수정
