+++ src/main/java/com/team9/anicare/community/controller/CommentController.java
@@ -1,6 +1,5 @@
 package com.team9.anicare.community.controller;
 
-import com.team9.anicare.common.Result;
 import com.team9.anicare.community.dto.CommentRequestDTO;
 import com.team9.anicare.community.dto.CommentResponseDTO;
 import com.team9.anicare.community.dto.LikeResponseDTO;
@@ -11,9 +10,11 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api/community")
+@RequestMapping("/api")
 public class CommentController {
 
     private final CommentService commentService;
@@ -56,6 +57,16 @@ public class CommentController {
     public ResponseEntity<LikeResponseDTO> createLike(Long userId, @PathVariable Long postingId) {
 
         LikeResponseDTO likeResponseDTO = commentService.createLike(userId, postingId);
+
         return ResponseEntity.status(HttpStatus.CREATED).body(likeResponseDTO);
+    }
+
+    @Operation(summary = "답글 조회")
+    @GetMapping("/comments/{parentId}/replies")
+    public ResponseEntity<List<CommentResponseDTO>> getReplies(Long userId, @PathVariable Long parentId) {
+
+        List<CommentResponseDTO> replies = commentService.getReplies(userId, parentId);
+
+        return ResponseEntity.ok(replies);
     }
 }
