+++ src/main/java/com/team9/anicare/community/model/Comment.java
@@ -30,8 +30,4 @@ public class Comment {
     @ManyToOne
     @JoinColumn(name = "parent_id")
     private Comment parent;
-
-    @OneToMany(mappedBy = "parent", cascade = CascadeType.ALL)
-    private List<Comment> replies;
-
 }
